Dijkstra optimization -- not backward compatible
Implemented a heap algorithm to calculate the Dijkstra path. Also made
backward-incompatible changes to the interface. Rather than evaluating
a single path, the cost is stored and accessed through a property of
the same name for a given start point. That way, multiple paths can be
investigated without having to recalculate, as long as the start point
remains invariant.

Evaluation is tied to the start point:

    - changing the grid through DijkstraPath.grid = grid invalidates
the start point and resets cost_function to the default.
    - changing the cost function, if start is set, sparks an immediate
recalculation. Set start to None to avoid recalculation.
    - Changing the start point sparks an immediate recalculation.